+++ hansen-cli/packages/template-react/CHANGELOG.md
@@ -1,7 +1,13 @@
 <!--
+
+## 1.2.0
+### Minor Changes
+
+- 修改 tempalte 的 name 为 projectName
  * @Date: 2024-12-23 18:03:57
  * @Description: description
 -->
+
 # @hansen-cli/template-react
 
 ## 1.1.0