+++ hansen-cli/packages/template-vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hansen-cli/template-vue
 
+## 1.2.0
+
+### Minor Changes
+
+- 修改 tempalte 的 name 为 projectName
+
 ## 1.1.0
 
 ### Minor Changes